Update Skia expectation for pixel test
Pixel test radial_shading_point_at_border.in contains a radial shading
of which one of the control point has a radius of 0. This causes
rendering discrepancies between multiple rendering tools (See [1]).

Even though Skia rendering result is different from AGG's expectation,
it does align with rendering results from mac Preview and Okular. In
addition to that, it aligns with Acrobat when the first control
point's radius is extremely small. Therefore, we keep Skia's rendering
behavior as it is and update its expectation.

[1] https://bugs.chromium.org/p/pdfium/issues/detail?id=1310#c12

Bug: pdfium:1310
Change-Id: I9729f7806fd4a534c65df11845c63a6091709616