Add HardClipScaled() to clip points relative to previous points
The vertices of lines are clamped before added to the path. Because of
this, lines can be drawn with an incorrect slope if the coordinate
value is clamped. Add HardClipScaled(), which will still clamp the
vertices, but also recalculates the x or y value so that the line
remains the same slope.

The original unit test's expected result was Skia's rendering, but there
is a minor difference between Skia's and AGG's rendering that is
almost unnoticeable. Add a new expected result for AGG.

Fixed: pdfium:1846
Change-Id: Ie96745d55b603b8440b4953310f9d00b2908765e