Add pixel test for image-path misalignment
Adds a pixel test that demonstrates misalignment between images and
paths with the same (fractional) coordinates. When coordinates do not
align exactly with pixels, PDFium uses different rounding logic for
these two cases, resulting in visually perceptible errors.

Both AGG and Skia backends currently get this wrong; the expected image
was generated by using a green rectangle in place of the green image in
the test file. When fixing this test, it's not important to match the
expected image exactly, just that the result is consistent between
images and paths.

Bug: pdfium:1973
Change-Id: Idff84fd7d76a1cfb01ca0aa25531e6255b54aea2