Change SkMipmapMode for smooth image rendering
Adds a test case bug_1396266.in that contains a masked image. When
Skia is enabled, the rendering result contains unwanted "dragged
shadows" instead of displaying recognizable numbers.

This Cl fixes the issue by changing SkMipmapMode from `kLinear` to
`kNearest` to achieve the best rendering result and avoid the dragging
effect.

Bug: chromium:1396266
Change-Id: I058228321d26241177b65336232a550e4f19f6d2