Check the first page xref size when parsing linearized PDFs.
When parsing linearized PDFs, the first page's cross-reference table
objects may not match up with the trailer's size entry. The current
behavior for handling PDFs with this error is to trust the trailer's
size entry, but that leads to incorrect PDF parsing, whereas other PDF
readers handle this error correctly.

Fix this issue by changing CPDF_Parser::StartLinearizedParse() to check
for the error. In which case, rebuild the cross-reference table.

To test this, add a deliberately bad bug_1287409.pdf. The steps to
create this PDF is as follows:

1) Run: qpdf --linearize rectangles.pdf bug_1287409.pdf
2) Deliberately change the trailer's /Size entry in a text editor.

The existing CPDF_Parser behavior results in an unknown /MediaBox for
the first page, and the page size will default to 612x792. With this
fix, the page size will be correctly set to 200x300.

Bug: chromium:1287409
Change-Id: I4f31a1775521ec43b02fb4bf88c75fe09f087815