Call a Skia initialization function from pdfium_test.
Since pdfium_test can be built in a Chromium checkout with a different
version of Skia, it needs to be able to do some initialization work for
that version of Skia. On the Chromium side, https://crrev.com/c/4196855
implements InitializeDiscardableMemoryAllocatorForTesting(). Here,
pdfium_test calls the function.

To make standalone PDFium builds works, provide a dummy
InitializeDiscardableMemoryAllocatorForTesting() in PDFium's Skia code.

Bug: chromium:1383711
Change-Id: I330e696f0298859d4b72d11577443202b543dec6