Specify output name for pdfium build component
Starting with chromium:r1096860, the Chromium build rules can modify the
resulting name of the built component.  In places where loading the
pdfium component library is hardcoded, the options are:
- modify the name in the source code where it is loaded, to follow
  the new naming scheme;
- use output_name in PDFium build rules to explicitly specify the
  name of the library, so that Chromium won't modify its name.

Use of the output_name approach is easier to understand when seeing the
pdfium component name hard-coded in source.  It also protects such code
blocks from any further modifications to the build rule.

Bug: chromium:1410951
Change-Id: Ib9a2375c760d58a521815dc08f3b8c5d31d3f36a