Fix an issue that a 0-length round-cap path was not rendered
For a 0-length closed path, it should be rendered as a dot if it has
the round line cap type. And it should not be rendered if its line cap
type is butt or projecting square.

This CL introduce the following rendering behaviors:
[1] For AGG and Skia: a closed 0-length round-cap path with only one
    "move" operation now can be rendered as a dot.
[2] For Skia only: Fixed the behaviors on rendering closed 0-length
    paths of all line cap types. Now that the "Zero Length Solid line"
    section of fx/path/bug_0_length_line_expected.pdf can be rendered
    correctly.

Note: The "Zero Length Solid line" section of corpus test
bug_0_length_line_expected.pdf is not completely fixed for AGG.

Bug: pdfium:1983,pdfium:1812
Change-Id: Ib2b23f9e0159a06f5ef4fe600a28c5b0d7a06a87