Fix rendering of stencil mask painted with tiling patterns
The stencil mask is painted with two patterns:
- first: a single cell with page dimensions, only the top half
         is filled black
- second: same as the first, but the second half is filled black

This combination should give us an output where both rectangles
are black. Because of this bug in pdfium, only the second is
black. This is due to the render device used to draw pattern image
being initialized with RGB color and filled white.
- the first pattern with the stencil mask, produces an output
  where the first rect is black second is white
- the second pattern with the stencil mask, produces an output
  where first rect is white second is black (overriding the previous
  pattern output)

Bug: pdfium:1966
Change-Id: I97fa5e30640d7a0d1d7e1970c2fbe96fcb3a85c2