Improve --regenerate_expected
Improves --regenerate_expected to regenerate the most specific expected
images that exist, which is what we want in most cases. In the rare case
that we want to regenerate a less specific expectation, we should delete
the more specific expectation first: Otherwise, we leave stale,
unnecessary expectations behind.

Note that this eliminates the need to specify whether we're regenerating
"all" or "platform" expectations, as any existing "platform" expectation
is more specific than the "all" expectation.

This requires treating the expectation order as a tree: Rather than AGG
being a strict, linear subset of Skia, AGG and Skia should have
branching expectations, with a common base of generic expectations.

This change also makes additional improvements to pngdiffer.py:

1. Fixes error reporting when expected files are missing.
2. Checks for "optipng" not just on Linux, since Regenerate() tries to
   execute it on all platforms anyway.
3. Prefixes more module-internal symbols with "_".
4. Migrates from distutils, which will be removed in Python 3.12.

This change also improves argparse usage in test_runner.py:

1. Quotes string literals consistently, using single quotes.
2. Drops the "dest" parameter if it's the same as the default.

Fixed: pdfium:1508
Change-Id: I39aa77d198a513fe92ea47f70aa6381b42ecce44