Update the CFX_FillRenderOptions used for flushing paths
Currently, CFX_SkiaDeviceDriver::PaintStroke() sets SkPaint based on
the most up-to-date CFX_FillRenderOptions `m_FillOptions`. But this
CFX_FillRenderOptions is not the same as the cached fill options
`m_fillOptions` in SkiaState when multiple paths with different
CFX_FillRenderOptions need drawing.

When CFX_SkiaDeviceDriver::PaintStroke() is called by
SkiaState::FlushPath(), it should set the SkPaint based on
`m_fillOptions` in the cached SkiaState, since FlushPath() is trying
to draw the already cached commands.

This CL makes the following improvements:
- Check if fill options have changed when determine when to flush
  cached paths.
- Make CFX_SkiaDeviceDriver::PaintStroke() take an
  CFX_FillRenderOptions as the input, so that the correct fill options
  can be used depending on the call site.

Change-Id: I120071431969d494afee7a0237d0b90273541527