Drop 11.pdf from the suppression list.
The Skia rendering behavior is correct for this test, which is drawing
black hairlines and then drawing white skinny images on top of them
inside the cirlce path.

There is no rule that the hairlines must be covered by the white
images and the rendering quality is controlled by how Skia handles
extremely "thin" image rendering internally. Therefore, update the
Skia expectation and drop this test from the suppression list.

Fixed: pdfium:1811
Change-Id: I6b5727c04332d7fc2c74163ec6b6eacb7f2a2339