Avoid creating JPX decoder when getting bad component value of 2.
When a image uses JPXDecode, a component value of 2 is a bad value
that doesn't match any color format. Most PDF rendering tools choose
to not render such image or report errors.

To be consistent with other PDF rendering tools, we should not create
JPX decoders for such images or rendering them.

Bug: pdfium:1747
Change-Id: I2c74ef1e15b4cc8353fd5b23de7eb10eca140e64