Avoid creating JPX decoder when getting bad component value.
When a PDF file doesn't specify the matching colorspace for images, it
will only support baseline color spaces and CMYK (The component number
will be 1,3 or 4). If the embedded JPX image has a unsupported
component number, most PDF rendering tools choose to not render such
images or report errors.

To be consistent with other PDF rendering tools, we should not create
JPX decoders for such images or render them.

Bug: pdfium:1747
Change-Id: I2c74ef1e15b4cc8353fd5b23de7eb10eca140e64