Move CXFA_EventParam::m_pTarget into CFXJSE_Engine
CXFA_EventParam can't be marked CPPGC_STACK_ALLOCATED(), because
CFXJSE_Engine -- which is not CPPGC_STACK_ALLOCATED() -- holds a
pointer to it. The only reason for the CPPGC_STACK_ALLOCATED()
annotation is to avoid needing to trace m_pTarget. With the field
removed, CXFA_EventParam no longer needs CPPGC_STACK_ALLOCATED().

This also adds CFXFSE_Engine::EventParamScope, which helps ensure that
event params and target and pushed and popped correctly.

Change-Id: I67e6abe08bb164051f47d30592ac1b0d027c4fa5