Get code coverage for CPDFSDK_Widget::Synchronize()
These tests require a mixed XFA form (i.e. no <config> XFA section,
say) with a PDF widget not overlapped by an XFA widget, and an XFA
widget linked by the same name to the PDF widget.

Send a click then type into the PDF widget to force it to sync
with the XFA one.

Change-Id: I71c5b346770b7da3045322b6af5765703f188556