Add fallback color space check when choosing JpxDecodeAction
When a JPEG2000 image object does not contain a /ColorSpace entry, we
rely on checking the JPEG2000 image's component number to determin
which JpxDecodeAction to take. However, the JPEG2000 image's color
component number is not a reliable factor to determine the color
channel number of the color space. A more reliable source would be
directly read the color space type from the JPEG2000 image's info.

Notice that Acrobat won't render JPEG2000 images if /ColorSpace or
/BitsPerComponent entry is missing, we do not necessarily have to
match Acrobat's behavior if we can determine color space type from the
image stream. Therefore, add the Skia expectation for bug_1258634.in
and remove it from the suppression list.

Bug: pdfium:1747
Change-Id: I2c74ef1e15b4cc8353fd5b23de7eb10eca140e64