Add minimal case for incorrect dashed lines on some Windows printers
Some Windows printers, for example "Microsoft Print to PDF" printer,
incorrectly print dashed lines. Add a minimal case to demonstrate this.

The minimal PDF sets a line width and sets the line cap to projecting
square cap. Chrome correctly renders the PDF with this line cap, but it
prints with a butt cap.

Bug: pdfium:1420150
Change-Id: If2e053f3b76116f99ae7f21c846491e0a452410e