Ensure {{include}} output ends on its own line
Ensures fixup_pdf_template.py ends the output of any {{include}}
directive on its own line. This ensures any lines after an {{include}}
are treated separately from the inclusion itself.

For example, macOS Preview doesn't treat bug_1469.jp2 as a valid JP2
file, because it has an extraneous newline character at the end. This
character previously was necessary to separate the inclusion from the
following "endstream" token. With the fix, the extra character in the
JP2 file is no longer needed.

Bug: pdfium:2000
Change-Id: I23963ee8ebe92db659640f0878d2f8cc019b68d3