Add basic /JPXDecode test cases
Adds test cases for the /JPXDecode filter that exercise all the basic
color spaces (gray, RGB, and CMYK), both with and without opacity. The
test file renders correctly in Adobe Acrobat Reader, but has issues in
most other PDF readers. The expectation file was created manually, based
on the PDF specification, and may not reflect our final implementation.

The files are included into the .in template using the {{include}}
mechanism, without additional filters. The gray and RGB files are in the
basic JP2 (.jp2) format, while the CMYK files are in JPX (.jpf) format,
as CMYK support is an extension.

This change includes a simple utility script, strip_jp2_comments.py, for
removing comment segments from a JPEG 2000 codestream.

Bug: pdfium:2000
Change-Id: I246ab1573b15089af9f2812d62b49ae67680d249