[Skia] Remove DebugVerify*IsPreMultiplied() calls
Removes DebugVerifyBitmapIsPreMultiplied() and
DebugVerifyDeviceIsPreMultiplied() calls, which act as assertions that a
bitmap or device has correct premultiplied alpha.

Some reasons to remove these calls:

1. They're very expensive (98% of CPU time in one profile), causing
   debug builds to differ dramatically from release builds in terms of
   performance.
2. Pixel tests provide cheaper, end-to-end validation.

The primary reason to keep these calls is for debugging specific alpha
calculations, but it doesn't make sense to slow down all debug builds
just for that.

Bug: pdfium:1949
Change-Id: Ia7d865af3f86d1854c8c39bccd84ac39e65275cf