Add MultiplyAlpha() API to CFX_RenderDevice
Adds a MultiplyAlpha() API to CFX_RenderDevice and
RenderDeviceDriverIface, and migrates
CPDF_RenderStatus::ProcessTransparency() to the new API. This allows AGG
and Skia to have different implementations.

For now, AGG and Skia both delegate to the corresponding CFX_DIBitmap
API. This API doesn't need to be implemented for the other driver types.

Also removes the unused CFX_SkiaDeviceDriver::GetBuffer() method.

Bug: pdfium:1949, pdfium:2007
Change-Id: Iad87c773d4af26786b6b6fb6bca057420731d493