Add fallback color space check when choosing JpxDecodeAction
The component number in JPX image info and the component number in a
PDF's color space are two different concepts. The component number
from a JPX image info might have included extra info besides color
channels. Therefore when trying to determine the correct decode action
or bitmap format to use, we should rely on the color space type from
PDF entries and the JPX image info instead of the the component number
from the JPX image info.

This CL implements the following improvements:
(1) When the a /ColorSpace entry is available in the PDF, use that PDF
    colorspace component number and JPX image color space type to
    determine the decode action.
(2) When /ColorSpace entry is not available, use JPX image color space
    type to determine the decode action.
(3) Since the component number in JPX image info doesn't stand for the
    correct color channel number, pass the correct color channel
    number into CJPX_Decoder::Decode().

Bug: pdfium:1747
Change-Id: I2c74ef1e15b4cc8353fd5b23de7eb10eca140e64