Check for removed resources in saved output in RemoveTextObject test
FPDFEditEmbedderTest.RemoveTextObject removes a page object, so the
resources associated with only that page object are not needed when
saving the PDF. Add checks to reflect the current situation where they
still exist, with a TODO to remove them.

Along the way, add an extra check to ensure the removed object is indeed
a text object.

Bug: chromium:1428724
Change-Id: Ia934158b5fd72a42fba0125aa1637d1c980bda3c