Update pdfium_test to output PDF
Extend PDF sample test to be able to output PDF documents of page
images.  This relies upon the pre-existing support for generating Skia
SKP format for pages.  This format is then run through Skia again to
be converted into PDF format.

This can be used as a proof of concept to see the output behavior from
trying to regenerate PDFs.

Bug: chromium:550205
Change-Id: I1596d3f7635294395fac8448014ce535f6e16efc