Update pdfium_test to output XPS
Extend PDF sample test to be able to output XPS documents of page
images.  This relies upon the pre-existing support for generating Skia
SKP format for pages.  This format is then run through Skia again to
be converted into XPS format.

This can be used as a proof of concept to see the output behavior from
converting PDFs into XPS format, which is of interest for using the XPS
print API on Windows.

Bug: chromium:1034053
Change-Id: I916b6f82c8e39427de0de6fe9a2709cf0e6cdf88