Save the trailer's object number when parsing
CPDF_CrossRefTable stores a CPDF_Dictionary for the trailer dictionary
object. For V5 trailers, the dictionary may be part of a stream object.
Since the dictionary is inline, the CPDF_Dictionary that represents it
does not know the object number. store this object number in
CPDF_CrossRefTable, alongside the CPDF_Dictionary, and expose it in a
pass-through method in CPDF_Parser.

Having this object number will make PDF object tree traversal return
more accurate results in an upcoming CL.

Bug: chromium:1428724,pdfium:2012
Change-Id: I86f980a09d2214c50412ce65a905dd92ebc85a6a