Add object tree traversal utility functions
Add GetObjectsWithMultipleReferences() to figure out which objects have
multiple references. Then code that manipulates PDFs can use this
function to determine if it is safe to edit an existing object. If it is
unsafe, the code can clone before editing instead.

Also add GetObjectsWithReference() to figure out which objects are
referenced. This can be used to determine which objects have no
references and can therefore be deleted.

Bug: chromium:1428724,pdfium:2012,pdfium:1409
Change-Id: I28817068d50c79f36d12bd50c1910937241194a5