Update pdfium_test filename generation to give helpful message
When pdfium_test needs to output filenames, those names have to be
restricted to a maximum length.  If this length is exceeded then it is
helpful to know the full string name which was too long.  This can't be
provided if the name is being built into a fixed-size array of the
maximum length.

Change the generation to instead use streaming to generate a string
which doesn't have a preset length limitation.  This allows for an
error message about the name to be provided in full.

Change-Id: I65ee1e8688b48c6af673f2ea2f9bbf7c32a0d344