Avoid rendering JPX image with mismatched PDF color space entry
PDF reference (ISO 32000-1:2008) page 89 mentioned that if color space
is present in the PDF, the color channels in the image data should
match the number of components in the PDF's color space.

Since the JPX image's channel number can be larger than PDF's
component number due to containing extra info, such as the
transparency level, we should make sure that the colorspace type from
the image align with that in the PDF entry. If the two color space
types do not match, such image should not be rendered.

Bug: chromium:1429465
Change-Id: Ia7d4ef9f05d017739308b5138df1939d08ce62ef