Fix roll-freetype.sh on Linux
Changes "sed -i ''" to "sed -i''" (removing the extra space), matching
the Chromium version of the roll-freetype.sh script. This fixes an
unintended difference in behavior between Linux and macOS.

On Linux, GNU sed treats the first version as "-i" and an (empty) input
filename, while it treats the second version as the (intended) "-i" with
an empty backup suffix. On macOS, the first version behaves like the
second version. (The extra separating space is not significant.)

Change-Id: I02566e4ccab0927fefc8ac1b995bb45f331e5ab3