Fix roll-freetype.sh on Linux
Changes "sed -i ''" to "sed -i~" (switching to a non-empty backup
suffix), and then delete the backup file with an explicit "rm" command.
This fixes an unintended difference in behavior between Linux and macOS.

On Linux, GNU sed treats the first version as "-i" and an (empty) input
filename, while it treats the second version as the "-i" with a "~"
suffix. On macOS, sed treats the first version as "-i" with an empty
suffix, while it treats the second version the same as GNU sed.

The main downside of the new approach is the extra backup file, which
must be removed after the "sed" command.

Change-Id: I02566e4ccab0927fefc8ac1b995bb45f331e5ab3