Use "Chrome Symbol" font on all platforms aside from Windows
Currently in CFX_FontMapper::FindSubstFont(), #if logic lets Apple
platforms and Android use an internal Symbol font when FindSubstFont()
calls for a Symbol font. The reasoning is that these platforms may not
provide a Symbol font. The same reasoning applies to Linux, so change
the #if condition to all platforms aside from Windows.

Bug: chromium:1442723
Change-Id: I1505fb45060fbe2293b6af976489d2878fa7eea0