Remove FT_Done_MM_Var() workaround
FT_Done_MM_Var() has been around since FreeType 2.9 from early 2018. Use
it directly to simplify code instead of assuming that it may not exist.
The default build, which uses a bundled FreeType, is up-to-date, so the
issue of FT_Done_MM_Var() being missing could only happen if someone is
building the code with an old system FreeType.

Bug: pdfium:1400
Change-Id: I7642de8884969484ff90a4cf865acc67e07e419b