Fix pdfium::span<> tests under BRP's asan_use_unowned_ptr=true
In this mode, there is not an option to discard a bad pointer, so
instead avoid making one in a common case. Since PDFium doesn't have
unit tests for spans (base should cover this before we roll in a new
version modulo our modifications), the failures manifest in tests
ByteStringView.FromVector and WideStringView.FromVector when we
make a string from a vector that has been cleared (no elements but
still has a backing store).

Bug: pdfium:2038
Change-Id: I856d72286aa55ef3e98629792564ede361188523