Avoid setting DiscardableMemoryAllocator in `LLVMFuzzerTestOneInput()`
A DiscardableMemoryAllocator is already set by `LLVMFuzzerInitialize()`
before calling `LLVMFuzzerTestOneInput()` when running `pdfium_fuzzer`,
which causes DCHECK failure in
DiscardableMemoryAllocator::SetInstance() since the allocator is already
set.

This CL removes the steps of setting up DiscardableMemoryAllocator in
`LLVMFuzzerTestOneInput()` to avoid the DCHECK failure. And it
simplifies the fuzzer process to only set up the disallocator once.

Bug: 1440491
Change-Id: I48e678253dacd696bd0fbf01805e150d9026ac53