Add CFX_SkiaDeviceDriver(SkCanvas*) constructor
Adds a new CFX_SkiaDeviceDriver constructor that accepts an (unowned)
SkCanvas. This allows CFX_SkiaDeviceDriver to use any SkCanvas
implementation, rather than just the ones backed by SkBitmap and
SkPictureRecorder.

Removes the constructor accepting an unowned SkPictureRecorder, as this
is equivalent to passing the result of getRecordingCanvas() to the
constructor accepting SkCanvas.

For the constructor accepting a CFX_DIBitmap, replaces the SkCanvas
wrapping an SkBitmap with the SkCanvas associated with a raster
SkSurface. This makes the SkCanvas uniformly unowned, eliminating a
special case in the destructor.

Bug: pdfium:2034
Change-Id: I2fb4420807fd3a763b59bde889758d4fa52ab3e3