Limit the number of form XObjects within a form XObject
Since a form XObject can contain other form XObjects, it is not too
difficult to construct one that contains M^N objects, where M is the
number of direct descendents and N is the depth of the object tree.
Since this just involves a large number of objects and there is no loop
in the object tree, the existing abuse prevention code in the parser
does not try to stop this.

To detect this issue, update CPDF_Form::RecursionState to record the
form count. Then record the number of form XObjects within a form
XObject in CPDF_StreamContentParser, and bail out once the limit has
been reached. The limit is set to several times higher than the highest
form count seen when parsing a large PDF corpus.

Bug: pdfium:1815
Change-Id: I671fb9578a0215fd32543a378ae9268d854bba81