Make CFX_DIBBase::GetBuffer() const-correct
Changes CFX_DIBBase::GetBuffer() to return a read-only view of the
buffer (a pdfium::span<const uint8_t>), rather than a writable view of
the buffer (a pdfium::span<uint8_t>). This prevents writing to the pixel
memory of a const CFX_DIBBase.

Callers that need a writable view of the buffer must now call
GetWritableBuffer() instead, which requires a non-const CFX_DIBBase.

Bug: pdfium:2034
Change-Id: I2f2055559c08c696c5ae0cbc920f3b7714487811