Fix PDF saving when multiple graphics states are in effect
To avoid saving unnecessary data, CL [1] added the ability to keep track
of which ExtGState resources were in use. This does not work in all
cases, like the newly created multiple_graphics_states.pdf being added
in this CL. To fix this case where multiple graphics states are in
effect at the same time, change the code that keeps track of ExtGState
resources to store a vector of resource names, instead of just the most
recently used one.

[1] https://pdfium-review.googlesource.com/105614

Bug: chromium:1451953
Change-Id: I673fe9f13efe90c50dbfa2511ab57092a2f987e3