[Skia] Handle text scaling inside CFX_SkiaDeviceDriver
Adds text scaling process in CFX_SkiaDeviceDriver when a glyph's
default width is different from the width assigned to it by the PDF
file. A glyph can be scaled down or up depending on its assigned
width.

This change enables Skia to handle more text rendering instead of
falling back to the common rendering paths as AGG (using DrawNormalTextHelper()).

Bug: pdfium:2053
Change-Id: I394a6faa226e4eef6319b91c514aae51ffc355ed