Tolerate bad /Size in V5 xrefs
When the /Size in a V5 xref is wrong, CPDF_Parser currently rejects xref
sections that have object numbers that goes over the incorrect size
limit. Other PDF implementations do not do this, and PDFium does not do
this for V4 xrefs. Change the code in CPDF_Parser::LoadCrossRefV5() that
is too strict and ignore the /Size.

Bug: chromium:1288804
Change-Id: I74d3f4ed548b71923c6e8b0c3948ff5defae3e8e