Fix stack-overflow in CheckForSharedFormInternal()
CheckForSharedFormInternal() is called recursively to traverse the
nodes in a meta data tree. If the meta data contains corrupted data
and never closes XML tags correctly, it could trigger stack-overflow
in this function.

This CL adds `kMaxMetaDataDepth` as the arbitrary maximum meta data
tree depth to prevent too many recursive calls of
CheckForSharedFormInternal() and adds a unit test for it. Also
renames `pElement` to `xml_element` to fit the c++ style guide better.

Bug: 1012377
Change-Id: Iaad7a07c6dc17d37130fdfd1e02441ba50b2066e