Add GDI bots
Adds "win_gdi" and "win_gdi_skia" bots that test GDI support on the CQ
and waterfall. (GDI is a Windows-specific technology. It also depends on
the default renderer, so it's important to test both with and without
Skia enabled.)

Also updates notify_emails to reflect the current team.

Fixed: pdfium:2054
Change-Id: Iaa752bfe15822feedd7251d408adfcb2a702f902