Replace FPDF_RECORDER with FPDF_SKIA_CANVAS
Replaces FPDF_RECORDER (an opaque pointer to an SkPictureRecorder) with
FPDF_SKIA_CANVAS (an opaque pointer to an SkCanvas). An SkCanvas is
strictly more general than an SkPictureRecorder.

Since we no longer need the definition of SkPictureRecorder, we can use
forward declarations more aggressively, and reduce the //fpdfsdk
target's dependency on //skia.

This change also reorganizes FPDF_RenderPageSkia() to follow the
structure of the other FPDF_RenderPage*() APIs more consistently.

Bug: pdfium:2059
Change-Id: If4855d61a39acc1f96e4c5adb9373bb4d26e19fe