Make "--diff" option handle reverse-byte-order results for pdfium_diff
Run pdfium_diff with "--diff" option can generate a .png file to
indicate the difference between the actual rendering result and the
baseline result. However this option currently does not support
comparing the reverse-byte-order rendering result to the baseline.

This CL adds the support to do such comparison so that it's easier for
users to compare the images.

Bug: pdfium:1955
Change-Id: I8b37975dfc902750722c82768489fefe66dc439b