Update `bPrinting` in RenderPageImpl()
In RenderPageImpl(), if the embedder is trying to render a PDF to be
displayed as the preview of the printing result, the rendering flags
will contain FPDF_PRINTING and the device type is kDisplay.

To make sure that such print preview matches the actual printing
result, `bPrinting` should not only depend on device type, but also be
set to true as long as (flags & FPDF_PRINTING) is true.

Bug: pdfium:1658
Change-Id: I49fb8e0527e609c842f55e49c11e05f649de0c65