Make generate_cas_paths.py output deterministic
Sorts final resolved paths in generate_cas_paths.py, as the order of
Path.iterdir() is not deterministic.

The //testing/tools:test_runner_py is a "leaf" target, so this doesn't
make any difference right now, but it could reduce rebuilds for future
targets that might depend on this target.

Bug: pdfium:1933
Change-Id: Ifb3029f27e48dd393cb203b7395bd837e269f384