Skip calling PatternValue::SetComps() when `values` is empty
PatternValue::SetComps() sometimes gets empty `values`. This is problematic when ubsan is enabled, which triggers runtime error that
fxcrt::spancpy() is copying from an null argument which should never
be null.

This CL skips calling PatternValue::SetComps() when `values` is empty
to avoid such error and adds a DCHECK() to prevent regression.

Bug: pdfium:1825
Change-Id: Iefbc7563dffb0d18bf9ab63813b16c88258aa90e