Avoid misalinged loads in LoadGlobalPersistentVariablesFromBuffer().
Chrome doesn't use any of this code, but the pdfium_unittests will
trip a warning about this under a full-up ubsan build. Fix this
just for the sake of quieting the bots.

Once the fixed header is extracted from the buffer, the remaining
data is packed and can't be expected to align on machine words.

Bug: pdfium:2066
Change-Id: Idc9edcef50cfee5899a1fb82c39a1299399382cc