[PA] Fix public_deps in pdfium.
The intermediate target should depend publicly on the destination to
forward the ability to include headers to its dependants. A few
public_deps were missing in pdfium.

See GN doc explaining this:
https://gn.googlesource.com/gn/+/master/docs/reference.md#discussion

Context: If we don't apply this patch, the PartitionAllocator library is
not going to be able to stand alone completely. Its public_config is not
going to be properly propagated toward all of its dependants. It will
fail to compile thme

Proofs:
- core/dpdfapi/page/BUILD.gn
  https://source.chromium.org/search?q=%23include.*parser%20file:.h&sq=&ss=chromium%2Fchromium%2Fsrc:third_party%2Fpdfium%2Fcore%2Ffpdfapi%2Fpage%2F
- core/fxcrt/BUILD.gn
  https://source.chromium.org/search?q=file:.h%20%23include.*partition_alloc&sq=&ss=chromium%2Fchromium%2Fsrc:third_party%2Fpdfium%2Fcore%2Ffxcrt%2F
- testing/BUILD.gn
  https://source.chromium.org/search?q=file:.h%20%23include.*fxcrt&sq=&ss=chromium%2Fchromium%2Fsrc:third_party%2Fpdfium%2Ftesting%2F

Bug: chromium:957519
Change-Id: I20a12818fc9d2b39ed9284f3fabc3c9e91ce0b1b