Reland "Don't include missing paths in CAS tree"
This is a reland of commit 16c574c18a2ad1530425106ea76b13ff9f8be4b2

Includes "data_deps" as "deps" for the "test_runner_py" target, as each
kind of dependency has a behavior this target needs: "data_deps" ensures
the dependencies are included as runtime dependencies, while "deps"
ensures the dependencies are built before the action runs. ("data_deps"
does not guarantee this.)

Tested this change on a clean build (which was breaking) and an unclean
build (to make sure nothing new has broken).

Original change's description:
> Don't include missing paths in CAS tree
>
> Skips paths that don't exist when archiving the CAS tree. If the missing
> paths aren't needed locally, they aren't needed in the CAS tree, either.
>
> Emits a warning in case we want to fix the incorrect runtime
> dependencies at the source.
>
> Bug: pdfium:1933
> Change-Id: Ib7ae7d710f73e98adc8d1fe6631ceef000efc6b5
> Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/110330
> Reviewed-by: Lei Zhang <thestig@chromium.org>
> Commit-Queue: K. Moon <kmoon@chromium.org>
> Commit-Queue: Lei Zhang <thestig@chromium.org>
> Auto-Submit: K. Moon <kmoon@chromium.org>

Bug: pdfium:1933
Change-Id: Ib910f5d41819c8f9aa8c26b273f3789004c5da48